pyproject.toml
[project]
name = "fleetsmart-mcp"
version = "0.1.0"
description = "FleetSmart MCP server for Claude"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
fleetsmart-mcp = "main:main"
